+++ helpers/general.js
@@ -208,9 +208,8 @@ export function createOpenOrDeleteNoteCallbackUrl(
   const paramStr = isLineLink ? 'noteTitle' : isFilename ? `filename` : paramType === 'date' ? `noteDate` : `noteTitle`
   const xcb = `noteplan://x-callback-url/${isDeleteNote ? 'deleteNote' : 'openNote'}?${paramStr}=`
   const head = heading && heading.length ? encodePlusParens(heading.replace('#', '')) : ''
-  // console.log(`createOpenOrDeleteNoteCallbackUrl: ${xcb}${titleOrFilename}${head ? `&heading=${head}` : ''}`)
   const encodedTitleOrFilename = encodePlusParens(titleOrFilename)
-  logDebug('createOpenOrDeleteNoteCallbackUrl', `encodePlusParens -> '${encodedTitleOrFilename}'`)
+  // logDebug('createOpenOrDeleteNoteCallbackUrl', `encodePlusParens -> '${encodedTitleOrFilename}'`)
   const openAs = openType && ['subWindow', 'splitView', 'useExistingSubWindow'].includes(openType) ? `&${openType}=yes` : ''
   let retVal = ''
   if (isLineLink) {